Improved serial spaceball compatibility
Some times, or some devices do not seem to send a CR before the first
initialization line (@1), and would not get picked up by spacenavd,
which was matching "\r@1". Removing the CR from the test should not
introduce false positives, as I don't think magellan devices send any
similar strings, so it should be safe to just look for "@1".

Also some devices seem to take a little too long to initialize and miss
our initialization string. Adding a 1second sleep after opening the
device seems to help.

Thanks to Benjamin Bergman for debugging and reporting this.